Trekker ut testutil. Rydder.
