+++ src/main/kotlin/no/nav/k9/los/nyoppgavestyring/mottak/oppgave/OppgaveV3Repository.kt
@@ -155,32 +155,6 @@ class OppgaveV3Repository(
         )
     }
 
-    @VisibleForTesting
-    fun hentAlleOppgaver(tx: TransactionalSession): List<OppgaveV3> {
-        return tx.run(
-            queryOf(
-                """
-                    select ot.ekstern_id as ot_ekstern_id, * from oppgave_v3 inner join oppgavetype ot on ot.id = oppgavetype_id where aktiv = true
-                """.trimIndent()
-            ).map { row ->
-                val oppgavetype =
-                    oppgavetypeRepository.hentOppgavetype(row.string("kildeomrade"), row.string("ot_ekstern_id"))
-                OppgaveV3(
-                    id = OppgaveId(row.long("id")),
-                    eksternId = row.string("ekstern_id"),
-                    eksternVersjon = row.string("ekstern_versjon"),
-                    oppgavetype = oppgavetype,
-                    status = Oppgavestatus.valueOf(row.string("status")),
-                    endretTidspunkt = row.localDateTime("endret_tidspunkt"),
-                    kildeområde = row.string("kildeomrade"),
-                    reservasjonsnøkkel = row.stringOrNull("reservasjonsnokkel") ?: "mangler_historikkvask",
-                    aktiv = row.boolean("aktiv"),
-                    felter = hentFeltverdier(OppgaveId(row.long("id")), oppgavetype, tx)
-                )
-            }.asList
-        )
-    }
-
     fun hentEksternIdForOppgaverMedStatus(
         oppgavetype: Oppgavetype,
         område: Område,
